fix: new prop favorXPath in parser_configuration
Add new boolean property "favorXPath" in parser_configuration.json

In some cases, a page's content may include JSON-LD (sdJSON) but
the data (e.g. price, name) is outdated/wrong compared to what can
be scraped with ParserXPath (this is the case with products pages
on unrealengine.com/marketplace for example).
In this case, the property "favorXPath" can be set to "true" for
a specific domain in order to ignore the problematic JSON-LD data.

Closes #55